refactor(core): move fold-only and test-only code out of create.rs
